media: qcom: camss: Pass CAMSS subdev callbacks via resource ops pointer
It is possible to pass all of the CAMSS subdevice internal operations
pointers from the controlling resources structure with an additional
pointer added to the resources structure.

This allows for the removal of most of the probe-time control structures.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Change-Id: I6570253e8242421b6b73a84718010c72e23424c9
Signed-off-by: Atanas Filipov <afilipov@qti.qualcomm.com>
(cherry picked from commit c23c7998161595b19ee1162da09609c31b5b4393)